# build/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "@infavy/api-server",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/users/{uid}": {
      "get": {
        "summary": "Fetch user document by UID",
        "parameters": [
          {
            "in": "path",
            "name": "uid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "summary": "Update user displayName and/or photo",
        "parameters": [
          {
            "in": "path",
            "name": "uid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "displayName": {
                    "type": "string"
                  },
                  "photoBase64": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated"
          },
          "400": {
            "description": "No updates provided"
          }
        }
      },
      "delete": {
        "summary": "Archive and delete user",
        "parameters": [
          {
            "in": "path",
            "name": "uid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User archived and deleted"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
    "/api/v1/watchlist/list": {
      "get": {
        "summary": "Get watchlist for a user",
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
...
}